+++ packages/bluesky_cli/lib/src/commands/codegen/tools/ozone/moderation.dart
@@ -13,6 +13,7 @@ import 'package:args/command_runner.dart';
 // Project imports:
 import 'moderation/cancel_scheduled_actions.dart';
 import 'moderation/emit_event.dart';
+import 'moderation/get_account_preferences.dart';
 import 'moderation/get_account_timeline.dart';
 import 'moderation/get_event.dart';
 import 'moderation/get_record.dart';
@@ -35,6 +36,7 @@ final class ToolsOzoneModerationCommand extends Command<void> {
   ToolsOzoneModerationCommand() {
     addSubcommand(CancelScheduledActionsCommand());
     addSubcommand(EmitEventCommand());
+    addSubcommand(GetAccountPreferencesCommand());
     addSubcommand(GetAccountTimelineCommand());
     addSubcommand(GetEventCommand());
     addSubcommand(GetRecordCommand());